Approx 1 minute corrected
Since Ake is talking a lot in this conversation the time elapsed is greater than the time that Tangata (the Aitu speaker who is the target in this file) is talking. But approx 1 min of Tangata has been corrected at this stage. This is the first 171 seconds of the file.